sign casting of exponent

+++ 16-binary_tree_is_perfect.c
@@ -15,7 +15,7 @@ int binary_tree_is_perfect(const binary_tree_t *tree)
 	if (!tree)
 		return (0);
 
-	if (binary_tree_size(tree) ==  (1 >> binary_tree_height(tree)) - 1)
+	if (binary_tree_size(tree) ==  (1 >> (size_t)binary_tree_height(tree)) - 1)
 		return (1);
 
 	return (0);
